Updated pragmas for compatibility with OMP 2 (int vs. uint32_t)

+++ MatrixReglistIO.cpp
@@ -106,7 +106,7 @@ genemap sampleFromGlobalGenemap() {
 	std::vector<std::vector<float>> subsampled_vecs(global_gm.size(), std::vector<float>(tot_num_subsample));
 	// parallelized can modify a vector
 #pragma omp parallel for num_threads(nthreads)
-	for (unsigned long gene = 0; gene < subsampled_vecs.size(); ++gene) {
+	for (int gene = 0; gene < subsampled_vecs.size(); ++gene) {
 		const std::vector<float> &expr_vec = global_gm[gene];
 		for (uint16_t i = 0U; i < tot_num_subsample; ++i)
 			subsampled_vecs[gene][i] = expr_vec[fold[i]];
